Consolidate group modules
