+++ modules/groups/components/List.tsx
@@ -6,8 +6,8 @@ import {
 } from "@mui/material";
 import React from "react";
 
-import { GroupFieldsFragment } from "../../graphql/generated";
-import GroupCard from "./Card";
+import Card from "./Card";
+import { GroupFieldsFragment } from "graphql/generated";
 
 interface Props {
   groups?: Array<GroupFieldsFragment | null | undefined> | null;
@@ -16,7 +16,7 @@ interface Props {
   paginationProps?: TablePaginationProps;
 }
 
-const GroupList: React.FC<Props> = ({
+const List: React.FC<Props> = ({
   groups,
   loading,
   toolbar,
@@ -29,9 +29,7 @@ const GroupList: React.FC<Props> = ({
       {groups?.length === 0 && (
         <Typography variant="h4">No groups found</Typography>
       )}
-      {groups?.map(
-        (group) => group && <GroupCard key={group.id} group={group} />
-      )}
+      {groups?.map((group) => group && <Card key={group.id} group={group} />)}
       {/* https://github.com/mui/material-ui/issues/15827 */}
       {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
       {/* @ts-ignore */}
@@ -44,4 +42,4 @@ const GroupList: React.FC<Props> = ({
   );
 };
 
-export default GroupList;
+export default List;
