+++ modules/groups/components/RoleChip.tsx
@@ -1,7 +1,7 @@
 import { Chip, Tooltip } from "@mui/material";
 import React from "react";
 
-import { GroupFieldsFragment } from "../../graphql/generated";
+import { GroupFieldsFragment } from "graphql/generated";
 import groupRoles from "utils/groupRoles";
 
 interface Props {
