+++ modules/groups/index.ts
@@ -1,0 +1,4 @@
+export { default as Browse } from "./browse";
+export { default as Group } from "./group";
+export { default as Manage } from "./manage";
+export { default as New } from "./new";
