+++ pages/groups/[id]/index.tsx
@@ -1,0 +1,7 @@
+import type { NextPage } from "next";
+
+import { Group } from "modules/groups";
+
+const GroupPage: NextPage = () => <Group />;
+
+export default GroupPage;
