+++ pages/groups/new.tsx
@@ -1,25 +1,7 @@
-import { Container, Typography } from "@mui/material";
-import { useRouter } from "next/router";
-import { useSnackbar } from "notistack";
-import React from "react";
+import type { NextPage } from "next";
 
-import GroupForm, { SubmitCallback } from "../../components/groups/GroupForm";
+import { New } from "modules/groups";
 
-const CreateGroup = () => {
-  const { enqueueSnackbar } = useSnackbar();
-  const router = useRouter();
+const NewGroup: NextPage = () => <New />;
 
-  const onSubmit: SubmitCallback = async ({ name }) => {
-    enqueueSnackbar(`Group ${name} created.`, { variant: "success" });
-    router.push("/groups");
-  };
-
-  return (
-    <Container>
-      <Typography variant="h4">Create a new group</Typography>
-      <GroupForm onSubmit={onSubmit} buttonLabel="Create" />
-    </Container>
-  );
-};
-
-export default CreateGroup;
+export default NewGroup;
